% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_matrix.R
\docType{methods}
\name{Arith,gpuMatrix,gpuMatrix-method}
\alias{Arith,gpuMatrix,gpuMatrix-method}
\title{gpuMatrix Arith methods}
\usage{
\S4method{Arith}{gpuMatrix,gpuMatrix}(e1, e2)
}
\arguments{
\item{e1}{A gpuMatrix object}

\item{e2}{A gpuMatrix object}
}
\description{
gpuMatrix Arith methods
}

