% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{Arith,gpuMatrix,missing-method}
\alias{Arith,gpuMatrix,missing-method}
\title{gpuMatrix Arith methods}
\usage{
\S4method{Arith}{gpuMatrix,missing}(e1, e2)
}
\arguments{
\item{e1}{A gpuMatrix object}

\item{e2}{missing}
}
\value{
A gpuMatrix object
}

