% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{Arith,gpuMatrix,numeric-method}
\alias{Arith,gpuMatrix,numeric-method}
\title{gpuMatrix Arith methods}
\usage{
\S4method{Arith}{gpuMatrix,numeric}(e1, e2)
}
\arguments{
\item{e1}{A gpuMatrix object}

\item{e2}{A numeric object}
}
\value{
A gpuMatrix object
}

