% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Arith,gpuVector,missing-method}
\alias{Arith,gpuVector,missing-method}
\title{gpuVector Arith methods}
\usage{
\S4method{Arith}{gpuVector,missing}(e1, e2)
}
\arguments{
\item{e1}{A gpuVector object}

\item{e2}{missing}
}
\value{
A gpuVector object
}

