% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Arith,gpuVector,numeric-method}
\alias{Arith,gpuVector,numeric-method}
\title{gpuVector Arith methods}
\usage{
\S4method{Arith}{gpuVector,numeric}(e1, e2)
}
\arguments{
\item{e1}{A gpuVector object}

\item{e2}{A numeric object}
}
\value{
A gpuVector object
}

