% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Arith,gpuVectorSlice,gpuVectorSlice-method}
\alias{Arith,gpuVectorSlice,gpuVectorSlice-method}
\title{gpuVector Arith methods}
\usage{
\S4method{Arith}{gpuVectorSlice,gpuVectorSlice}(e1, e2)
}
\arguments{
\item{e1}{A gpuVectorSlice object}

\item{e2}{A gpuVectorSlice object}
}
\value{
A gpuVector object
}

