% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_vector.R
\docType{methods}
\name{Arith,igpuVector,igpuVector-method}
\alias{Arith,igpuVector,igpuVector-method}
\title{gpuVector Arith methods}
\usage{
\S4method{Arith}{igpuVector,igpuVector}(e1, e2)
}
\arguments{
\item{e1}{A igpuVector object}

\item{e2}{A igpuVector object}
}
\description{
gpuVector Arith methods
}

