% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{Arith,numeric,gpuMatrix-method}
\alias{Arith,numeric,gpuMatrix-method}
\title{gpuMatrix Arith methods}
\usage{
\S4method{Arith}{numeric,gpuMatrix}(e1, e2)
}
\arguments{
\item{e1}{A numeric object}

\item{e2}{A gpuMatrix object}
}
\value{
A gpuMatrix object
}

