% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{Arith,numeric,vclVector-method}
\alias{Arith,numeric,vclVector-method}
\title{vclVector Arith methods}
\usage{
\S4method{Arith}{numeric,vclVector}(e1, e2)
}
\arguments{
\item{e1}{A numeric object}

\item{e2}{A vclVector object}
}
\value{
A vclVector object
}

