% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{Arith,vclMatrix,missing-method}
\alias{Arith,vclMatrix,missing-method}
\title{vclMatrix Arith methods}
\usage{
\S4method{Arith}{vclMatrix,missing}(e1, e2)
}
\arguments{
\item{e1}{A vclMatrix object}

\item{e2}{missing}
}
\value{
A vclMatrix object
}

