% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{Arith,vclMatrix,vclMatrix-method}
\alias{Arith,vclMatrix,vclMatrix-method}
\title{vclMatrix Arith methods}
\usage{
\S4method{Arith}{vclMatrix,vclMatrix}(e1, e2)
}
\arguments{
\item{e1}{A vclMatrix object}

\item{e2}{A vclMatrix object}
}
\value{
A vclMatrix object
}

