% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{Arith,vclVector,missing-method}
\alias{Arith,vclVector,missing-method}
\title{vclVector Arith methods}
\usage{
\S4method{Arith}{vclVector,missing}(e1, e2)
}
\arguments{
\item{e1}{A vclVector object}

\item{e2}{missing}
}
\value{
A vclVector object
}

