% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{Arith,vclVector,numeric-method}
\alias{Arith,vclVector,numeric-method}
\title{gpuVector Arith methods}
\usage{
\S4method{Arith}{vclVector,numeric}(e1, e2)
}
\arguments{
\item{e1}{A gpuVector object}

\item{e2}{A numeric object}
}
\value{
A gpuVector object
}

