% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\docType{methods}
\name{as.gpuMatrix}
\alias{as.gpuMatrix}
\title{Convert object to a gpuMatrix}
\usage{
as.gpuMatrix(object, type)
}
\arguments{
\item{object}{An object that is or can be converted to a 
\code{matrix} object}

\item{type}{A character string specifying the type of gpuMatrix.  Default
is NULL where type is inherited from the source data type.}
}
\value{
A gpuMatrix object
}
\description{
Construct a gpuMatrix of a class that inherits
from \code{gpuMatrix}
}
\author{
Charles Determan Jr.
}
