% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuSparseMatrix.R

\name{as.matrix.gpuSparseMatrix}
\alias{as.matrix.gpuSparseMatrix}
\title{Convert GPU Sparse Matrix to Dense Matrix}
\usage{
\method{as.matrix}{gpuSparseMatrix}(x, ...)
}
\arguments{
\item{x}{A gpuSparseMatrix object}

\item{...}{Additional arguments (for compatibility)}
}
\value{
A dense R matrix
}
\description{
Convert a GPU sparse matrix to a dense R matrix
}
