% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods-gpuVector.R
\docType{methods}
\name{as.gpuVector}
\alias{as.gpuVector}
\alias{as.gpuVector,vector-method}
\alias{as.gpuVector,vector}
\title{Convert object to a gpuVector}
\usage{
as.gpuVector(object, type)

\S4method{as.gpuVector}{vector}(object, type = NULL)
}
\arguments{
\item{object}{An object that is or can be converted to a 
\code{vector} object}

\item{type}{A character string specifying the type of gpuVector.  Default
is NULL where type is inherited from the source data type.}
}
\value{
A gpuVector object
}
\description{
Construct a gpuVector of a class that inherits
from \code{gpuVector}
}
\author{
Charles Determan Jr.
}
