% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asserts.R
\name{assert_has_double}
\alias{assert_has_double}
\title{Does device have 'double' support?}
\usage{
assert_has_double(device_idx, context_idx,
  severity = getOption("assertive.severity", "stop"))
}
\arguments{
\item{device_idx}{An integer indicating which device to query}

\item{context_idx}{An integer indicating which context to query}

\item{severity}{How severe should the consequences of the assertion be?}
}
\value{
Returns nothing but throws an error if device does not support
double precision
}
\description{
Function to query if device (identified by index) supports
double precision
}
\seealso{
\link{deviceHasDouble}
}
\author{
Charles Determan Jr.
}
