% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuSparseMatrix.R

\name{cgpuSparseMatrix-class}
\alias{cgpuSparseMatrix-class}
\alias{cgpuSparseMatrix}
\title{cgpuSparseMatrix Class}
\description{
A complex float type sparse matrix in the S4 \code{gpuSparseMatrix}
representation using Compressed Sparse Row (CSR) format.
}
\section{Slots}{

\describe{
    \item{\code{address}:}{Pointer to a complex float sparse matrix.}
}
}

\value{
If the gpuSparseMatrix object is of type 'complex float', returns TRUE, if not, returns an error message.
}

