% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{colnames}
\alias{colnames}
\alias{colnames.default}
\alias{colnames.gpuMatrix}
\alias{colnames.vclMatrix}
\alias{colnames<-,gpuMatrix-method}
\alias{colnames<-,vclMatrix-method}
\title{Row and Column Names}
\usage{
colnames(x, do.NULL, prefix)

\method{colnames}{default}(x, do.NULL = TRUE, prefix = "col")

\method{colnames}{gpuMatrix}(x, ...)

\S4method{colnames}{gpuMatrix}(x) <- value

\method{colnames}{vclMatrix}(x, ...)

\S4method{colnames}{vclMatrix}(x) <- value
}
\arguments{
\item{x}{A gpuR matrix object}

\item{do.NULL}{logical. If \code{FALSE} names are NULL, names are created. 
(not currently used)}

\item{prefix}{for create names. (not currently used)}

\item{...}{Additional arguments}

\item{value}{A character vector to assign as row/column names}
}
\description{
Retrieve or set the row or column names of a gpuR matrix object
}

