% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Compare,gpuVector,vector-method}
\alias{Compare,gpuVector,vector-method}
\title{Compare gpuVector and vector elements}
\usage{
\S4method{Compare}{gpuVector,vector}(e1, e2)
}
\arguments{
\item{e1}{A gpuvector object}

\item{e2}{A Vector object}
}
\description{
Compare gpuVector and vector elements
}

