% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Compare,vector,gpuVector-method}
\alias{Compare,vector,gpuVector-method}
\alias{Compare-vector-gpuVector}
\alias{Compare,gpuVector,vector-method}
\alias{Compare-gpuVector-vector}
\title{Compare vector and gpuVector elements}
\usage{
\S4method{Compare}{vector,gpuVector}(e1, e2)

\S4method{Compare}{gpuVector,vector}(e1, e2)
}
\arguments{
\item{e1}{A vector/gpuVector object}

\item{e2}{A vector/gpuVector object}
}
\value{
A logical vector
}
\description{
Methods for comparison operators
}
\author{
Charles Determan Jr.
}
