% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Compare,vector,gpuVector-method}
\alias{Compare,vector,gpuVector-method}
\title{Compare vector and gpuVector elements}
\usage{
\S4method{Compare}{vector,gpuVector}(e1, e2)
}
\arguments{
\item{e1}{A vector object}

\item{e2}{A gpuVector object}
}
\description{
Compare vector and gpuVector elements
}

