% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R, R/methods.R
\docType{methods}
\name{cov,vclMatrix,missing,missing,missing-method}
\alias{cov,gpuMatrix,missing,missing,character-method}
\alias{cov,gpuMatrix,missing,missing,missing-method}
\alias{cov,vclMatrix,missing,missing,character-method}
\alias{cov,vclMatrix,missing,missing,missing-method}
\title{Covariance (gpuR)}
\usage{
\S4method{cov}{vclMatrix,missing,missing,missing}(x, y = NULL, use = NULL,
  method = "pearson")

\S4method{cov}{vclMatrix,missing,missing,character}(x, y = NULL, use = NULL,
  method = "pearson")

\S4method{cov}{gpuMatrix,missing,missing,missing}(x, y = NULL, use = NULL,
  method = "pearson")

\S4method{cov}{gpuMatrix,missing,missing,character}(x, y = NULL, use = NULL,
  method = "pearson")
}
\arguments{
\item{x}{A gpuR object}

\item{y}{Not used}

\item{use}{Not used}

\item{method}{Character string indicating with covariance to be computed.}
}
\value{
A gpuMatrix/vclMatrix containing the symmetric covariance values.
}
\description{
Compute covariance values
}
\author{
Charles Determan Jr.
}

