% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{currentContext}
\alias{currentContext}
\title{Current Context}
\usage{
currentContext()
}
\value{
An integer reflecting the context listed in \link{listContexts}
}
\description{
Get current context index
}
