% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{currentDevice}
\alias{currentDevice}
\title{Current Device Information}
\usage{
currentDevice()
}
\value{
list containing

\item{device}{Character string of device name}

\item{device_index}{Integer identifying device}

\item{device_type}{Character string identifying device type (e.g. gpu)}
}
\description{
Check current device information
}
