% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{currentPlatform}
\alias{currentPlatform}
\title{Return Current Platform}
\usage{
currentPlatform()
}
\value{
\item{platform}{Name of the current platform}

\item{platform_index}{Index of current platform}
}
\description{
Find out which platform is currently in use
}
\seealso{
\link{detectPlatforms}
}
