% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_opencl.R
\name{custom_opencl}
\alias{custom_opencl}
\title{Custom OpenCL Kernels}
\usage{
custom_opencl(kernel, cl_args, type)
}
\arguments{
\item{kernel}{path to OpenCL kernel file}

\item{cl_args}{A \code{data.frame} that contains argument definitions.
Provided by \code{\link{setup_opencl}}}

\item{type}{The precision on which the kernel is compiled.  Options include
\code{"int"}, \code{"float"}, and \code{"double"}}
}
\description{
Compile a custom function using a user provided OpenCL kernel
}
