% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclMatrix.R
\docType{class}
\name{cvclMatrix-class}
\alias{cvclMatrix-class}
\title{cvclMatrix Class}
\value{
If the vclMatrix object is of type 'complex float', returns TRUE, if not, returns an error message.
}
\description{
An complex float type matrix in the S4 \code{vclMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a complex float type matrix}
 }
}

\seealso{
\code{\link{vclMatrix-class}}, 
\code{\link{ivclMatrix-class}},
\code{\link{fvclMatrix-class}}
}
\author{
Charles Determan Jr.
}
