% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R, R/methods.R
\docType{methods}
\name{det,vclMatrix-method}
\alias{det,vclMatrix-method}
\alias{det,vclMatrix}
\alias{det,gpuMatrix-method}
\alias{det,vclMatrix}
\title{Calculate Determinant of a Matrix on GPU}
\usage{
\S4method{det}{vclMatrix}(x)

\S4method{det}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuR matrix object}
}
\value{
The determine of \code{x}
}
\description{
\code{det} calculates the determinant of a matrix.
}
\note{
This function uses an LU decomposition and the \code{det} 
function is simply a wrapper returning the determinant product
}
\author{
Charles Determan Jr.
}
