% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{detectGPUs}
\alias{detectGPUs}
\title{Detect Available GPUs}
\usage{
detectGPUs(platform_idx = 1L)
}
\arguments{
\item{platform_idx}{An integer value indicating which platform to query.}
}
\value{
An integer representing the number of available GPUs
}
\description{
Find out how many GPUs available
}
\seealso{
\link{detectPlatforms}
}

