% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{detectPlatforms}
\alias{detectPlatforms}
\title{Detect Number of Platforms}
\usage{
detectPlatforms()
}
\value{
An integer value representing the number of platforms available.
}
\description{
Find out how many OpenCL enabled platforms are available.
}
\seealso{
\link{detectGPUs}
}
