% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers-utils.R
\name{deviceHasDouble}
\alias{deviceHasDouble}
\title{Check GPU double precision support}
\usage{
deviceHasDouble(platform_idx = 1L, gpu_idx = 1L)
}
\arguments{
\item{platform_idx}{An integer value indicating which platform to query.}

\item{gpu_idx}{An integer value indicating which gpu to query.}
}
\value{
A boolean designating whether the device supports double precision
}
\description{
This function checks the GPU device extensions for the
variable cl_khr_fp64 which means the device supports double precision.
}
\seealso{
\link{gpuInfo}
}

