% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers-utils.R
\name{deviceType}
\alias{deviceType}
\title{Check device type}
\usage{
deviceType(device_idx = NULL, context_idx = currentContext())
}
\arguments{
\item{device_idx}{An integer value indicating which device to query.}

\item{context_idx}{An integer value indicating which context to query.}
}
\value{
A character string indicating the device type
}
\description{
Check what type a device is given platform and device indices
}
