% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{dgpuBigMatrix-class}
\alias{dgpuBigMatrix-class}
\title{dgpuBigMatrix Class}
\description{
A double type matrix in the S4 \code{gpuBigMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{External pointer to shared
     double type memory mapped matrix.}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{gpuBigMatrix-class}},
\code{\link{igpuBigMatrix-class}},
\code{\link{fgpuBigMatrix-class}}
}

