% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuSparseMatrix.R

\name{dgpuSparseMatrix-class}
\alias{dgpuSparseMatrix-class}
\alias{dgpuSparseMatrix}
\title{dgpuSparseMatrix Class}
\description{
A double type sparse matrix in the S4 \code{gpuSparseMatrix}
representation using Compressed Sparse Row (CSR) format.
}
\section{Slots}{

\describe{
    \item{\code{address}:}{Pointer to a double type sparse matrix}
}
}

\value{
If the gpuSparseMatrix object is of type 'double', returns TRUE, if not, returns an error message.
}

