% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_matrix.R
\docType{methods}
\name{dim,gpuMatrix-method}
\alias{dim,gpuMatrix-method}
\title{gpuMatrix dim method}
\usage{
\S4method{dim}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\description{
gpuMatrix dim method
}

