% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuSparseMatrix.R

\name{dim-gpuSparseMatrix}
\alias{dim,gpuSparseMatrix-method}
\alias{nrow,gpuSparseMatrix-method}
\alias{ncol,gpuSparseMatrix-method}
\alias{nnz,gpuSparseMatrix-method}
\alias{typeof,gpuSparseMatrix-method}
\title{GPU Sparse Matrix Dimensions and Properties}
\usage{
\S4method{dim}{gpuSparseMatrix}(x)

\S4method{nrow}{gpuSparseMatrix}(x)

\S4method{ncol}{gpuSparseMatrix}(x)

\S4method{nnz}{gpuSparseMatrix}(x)

\S4method{typeof}{gpuSparseMatrix}(x)
}
\arguments{
\item{x}{A gpuSparseMatrix object}
}
\value{
For \code{dim}: A vector of length 2 with rows and columns

For \code{nrow}: Number of rows

For \code{ncol}: Number of columns

For \code{nnz}: The count of non-zero elements

For \code{typeof}: A character string indicating the data type
}
\description{
Get the dimensions and properties of a GPU sparse matrix
}

