% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{dim,vclMatrix-method}
\alias{dim,vclMatrix}
\alias{dim,vclMatrix-method}
\title{vclMatrix dim method}
\usage{
\S4method{dim}{vclMatrix}(x)
}
\arguments{
\item{x}{A vclMatrix object}
}
\value{
A length 2 vector of the number of rows and columns respectively.
}
\author{
Charles Determan Jr.
}

