% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-vclVector.R
\docType{class}
\name{dvclVector-class}
\alias{dvclVector-class}
\title{dvclVector Class}
\description{
An double vector in the S4 \code{vclVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double typed vector}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{vclVector-class}}
}

