% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{[,gpuVector,missing,missing,missing-method}
\alias{[,gpuVector}
\alias{[,gpuVector,missing,missing,missing-method}
\alias{[,gpuVector,numeric,missing,missing-method}
\alias{[<-,gpuVector,numeric,missing,integer-method}
\alias{[<-,gpuVector,numeric,missing,numeric-method}
\title{Extract/Set gpuVector elements}
\usage{
\S4method{[}{gpuVector,missing,missing,missing}(x, i, j, drop)

\S4method{[}{gpuVector,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{gpuVector,numeric,missing,numeric}(x, i, j) <- value

\S4method{[}{gpuVector,numeric,missing,integer}(x, i, j) <- value
}
\arguments{
\item{x}{A gpuVector object}

\item{i}{missing}

\item{j}{missing}

\item{drop}{missing}

\item{value}{data of similar type to be added to gpuMatrix object}
}
\description{
Extract/Set gpuVector elements
}
\author{
Charles Determan
}

