% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-gpuMatrix.R, R/extract-gpuVector.R,
%   R/extract-vclMatrix.R, R/extract-vclVector.R
\docType{methods}
\name{[,gpuMatrix,missing,missing,missing-method}
\alias{[,gpuMatrix,missing,missing,missing-method}
\alias{[,gpuMatrix,missing,numeric,missing-method}
\alias{[,gpuMatrix,numeric,missing,missing-method}
\alias{[,gpuMatrix,numeric,numeric,missing-method}
\alias{[<-,gpuMatrix,numeric,missing,numeric-method}
\alias{[<-,igpuMatrix,numeric,missing,integer-method}
\alias{[<-,gpuMatrix,missing,numeric,numeric-method}
\alias{[<-,igpuMatrix,missing,numeric,integer-method}
\alias{[<-,gpuMatrix,numeric,numeric,numeric-method}
\alias{[<-,igpuMatrix,numeric,numeric,integer-method}
\alias{[,gpuVector,missing,missing,missing-method}
\alias{[,gpuVector,numeric,missing,missing-method}
\alias{[<-,gpuVector,numeric,missing,numeric-method}
\alias{[<-,gpuVector,numeric,missing,integer-method}
\alias{[,vclMatrix,missing,missing,missing-method}
\alias{[,vclMatrix,missing,numeric,missing-method}
\alias{[,vclMatrix,numeric,missing,missing-method}
\alias{[,vclMatrix,numeric,numeric,missing-method}
\alias{[<-,vclMatrix,missing,numeric,numeric-method}
\alias{[<-,ivclMatrix,missing,numeric,integer-method}
\alias{[<-,vclMatrix,numeric,missing,numeric-method}
\alias{[<-,ivclMatrix,numeric,missing,integer-method}
\alias{[<-,vclMatrix,numeric,numeric,numeric-method}
\alias{[<-,ivclMatrix,numeric,numeric,integer-method}
\alias{[<-,vclMatrix,missing,missing,matrix-method}
\alias{[<-,vclMatrix,missing,missing,vclMatrix-method}
\alias{[<-,vclMatrix,missing,numeric,vclMatrix-method}
\alias{[<-,vclMatrix,missing,missing,numeric-method}
\alias{[<-,vclMatrix,missing,missing,vclVector-method}
\alias{[<-,vclMatrix,missing,numeric,vclVector-method}
\alias{[,vclVector,missing,missing,missing-method}
\alias{[,vclVector,numeric,missing,missing-method}
\alias{[<-,vclVector,numeric,missing,numeric-method}
\alias{[<-,ivclVector,numeric,missing,integer-method}
\alias{[<-,vclVector,logical,missing,numeric-method}
\alias{[<-,vclVector,missing,missing,numeric-method}
\alias{[<-,vclVector,missing,missing,vclVector-method}
\alias{[<-,vclVector,numeric,missing,vclVector-method}
\alias{[<-,vclVector,missing,missing,vclMatrix-method}
\alias{[<-,vclVector,numeric,missing,vclMatrix-method}
\title{Extract gpuR object elements}
\usage{
\S4method{[}{gpuMatrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,numeric,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{gpuMatrix,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,numeric,missing,numeric}(x, i, j, ...) <- value

\S4method{[}{igpuMatrix,numeric,missing,integer}(x, i, j) <- value

\S4method{[}{gpuMatrix,missing,numeric,numeric}(x, i, j) <- value

\S4method{[}{igpuMatrix,missing,numeric,integer}(x, i, j) <- value

\S4method{[}{gpuMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{igpuMatrix,numeric,numeric,integer}(x, i, j) <- value

\S4method{[}{gpuVector,missing,missing,missing}(x, i, j, drop)

\S4method{[}{gpuVector,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{gpuVector,numeric,missing,numeric}(x, i, j) <- value

\S4method{[}{gpuVector,numeric,missing,integer}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{vclMatrix,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{vclMatrix,numeric,missing,missing}(x, i, j, ..., drop = TRUE)

\S4method{[}{vclMatrix,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{vclMatrix,missing,numeric,numeric}(x, i, j) <- value

\S4method{[}{ivclMatrix,missing,numeric,integer}(x, i, j) <- value

\S4method{[}{vclMatrix,numeric,missing,numeric}(x, i, j, ...) <- value

\S4method{[}{ivclMatrix,numeric,missing,integer}(x, i, j) <- value

\S4method{[}{vclMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{ivclMatrix,numeric,numeric,integer}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,missing,matrix}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,missing,vclMatrix}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,numeric,vclMatrix}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,missing,numeric}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,missing,vclVector}(x, i, j) <- value

\S4method{[}{vclMatrix,missing,numeric,vclVector}(x, i, j) <- value

\S4method{[}{vclVector,missing,missing,missing}(x, i, j, drop)

\S4method{[}{vclVector,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{vclVector,numeric,missing,numeric}(x, i, j) <- value

\S4method{[}{ivclVector,numeric,missing,integer}(x, i, j) <- value

\S4method{[}{vclVector,logical,missing,numeric}(x, i, j) <- value

\S4method{[}{vclVector,missing,missing,numeric}(x, i, j) <- value

\S4method{[}{vclVector,missing,missing,vclVector}(x, i, j) <- value

\S4method{[}{vclVector,numeric,missing,vclVector}(x, i, j) <- value

\S4method{[}{vclVector,missing,missing,vclMatrix}(x, i, j) <- value

\S4method{[}{vclVector,numeric,missing,vclMatrix}(x, i, j) <- value
}
\arguments{
\item{x}{A gpuR object}

\item{i}{indices specifying rows}

\item{j}{indices specifying columns}

\item{drop}{missing}

\item{...}{Additional arguments}

\item{value}{data of similar type to be added to gpuMatrix object}
}
\description{
Operators to extract or replace elements
}
\author{
Charles Determan Jr.
}
