% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{[,vclMatrix,missing,missing,missing-method}
\alias{[,vclMatrix}
\alias{[,vclMatrix,missing,missing,missing-method}
\alias{[,vclMatrix,missing,numeric,missing-method}
\alias{[,vclMatrix,numeric,missing,missing-method}
\alias{[,vclMatrix,numeric,numeric,missing-method}
\title{Extract or Replace Parts of vclMatrix}
\usage{
\S4method{[}{vclMatrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{vclMatrix,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{vclMatrix,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{vclMatrix,numeric,numeric,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A vclMatrix object}

\item{i}{indices specifying rows}

\item{j}{indices specifying columns}

\item{drop}{missing}
}
\description{
Extract or Replace Parts of vclMatrix
}
\author{
Charles Determan Jr.
}

