% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{fgpuBigMatrix-class}
\alias{fgpuBigMatrix-class}
\title{fgpuBigMatrix Class}
\description{
A float type matrix in the S4 \code{gpuBigMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{External pointer to shared
     float type memory mapped matrix.}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{gpuBigMatrix-class}},
\code{\link{fgpuBigMatrix-class}},
\code{\link{dgpuBigMatrix-class}}
}

