% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuMatrix.R
\docType{class}
\name{fgpuMatrix-class}
\alias{fgpuMatrix-class}
\title{fgpuMatrix Class}
\description{
A float type matrix in the S4 \code{gpuMatrix}
representation.

An complex float type matrix in the S4 \code{gpuMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float matrix.}
 }


 \describe{
     \item{\code{address}:}{Pointer to a complex float matrix.}
 }
}
\author{
Charles Determan Jr.

Charles Determan Jr.
}
\seealso{
\code{\link{gpuMatrix-class}}, 
\code{\link{igpuMatrix-class}},
\code{\link{dgpuMatrix-class}}

\code{\link{gpuMatrix-class}}, 
\code{\link{igpuMatrix-class}},
\code{\link{dgpuMatrix-class}}
}

