% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuSparseMatrix.R

\name{fgpuSparseMatrix-class}
\alias{fgpuSparseMatrix-class}
\alias{fgpuSparseMatrix}
\title{fgpuSparseMatrix Class}
\description{
A float type sparse matrix in the S4 \code{gpuSparseMatrix}
representation using Compressed Sparse Row (CSR) format.
}
\section{Slots}{

\describe{
    \item{\code{address}:}{Pointer to a float typed sparse matrix.}
}
}

\value{
If the gpuSparseMatrix object is of type 'float', returns TRUE, if not, returns an error message.
}

