% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuVector.R
\docType{class}
\name{fgpuVector-class}
\alias{fgpuVector-class}
\title{fgpuVector Class}
\description{
An float vector in the S4 \code{gpuVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float typed vector}
 }
}

\seealso{
\code{\link{gpuVector-class}}
}
\author{
Charles Determan Jr.
}
