% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-vclMatrix.R
\docType{class}
\name{fvclMatrix-class}
\alias{fvclMatrix-class}
\title{fvclMatrix Class}
\description{
An integer type matrix in the S4 \code{vclMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float matrix.}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{vclMatrix-class}},
\code{\link{ivclMatrix-class}},
\code{\link{dvclMatrix-class}}
}

