% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclVector.R
\docType{class}
\name{fvclVector-class}
\alias{fvclVector-class}
\title{fvclVector Class}
\value{
If the vclVector object is of type 'float', returns TRUE, if not, returns an error message.
}
\description{
An float vector in the S4 \code{vclVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float typed vector}
 }
}

\seealso{
\code{\link{vclVector-class}}
}
\author{
Charles Determan Jr.
}
