% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{gpu_BigMat_mult}
\alias{gpu_BigMat_mult}
\title{GPU Big Matrix Multiplication}
\usage{
gpu_BigMat_mult(A, B)
}
\arguments{
\item{A}{A gpuBigMatrix object}

\item{B}{A gpuBigMatrix object}
}
\value{
A gpuBigMatrix object
}
\description{
big matrix gpu multiplication
}
\author{
Charles Determan Jr.
}

