% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{gpu_Mat_axpy}
\alias{gpu_Mat_axpy}
\title{GPU Matrix Multiplication}
\usage{
gpu_Mat_axpy(alpha, A, B)
}
\arguments{
\item{alpha}{Numeric value to multiply the A matrix}

\item{A}{A gpuMatrix object}

\item{B}{A gpuMatrix object}
}
\value{
A gpuMatrix object
}
\description{
matrix multiplication
}
\author{
Charles Determan Jr.
}

