% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{gpu_vec_subtr}
\alias{gpu_vec_subtr}
\title{GPU Vector Subtraction}
\usage{
gpu_vec_subtr(A, B)
}
\arguments{
\item{A}{A gpuVector object}

\item{B}{A gpuVector object}
}
\value{
A gpuVector object
}
\description{
vector subtraction
}
\author{
Charles Determan Jr.
}

