% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gpuBigMatrix-class}
\alias{gpuBigMatrix-class}
\title{gpuBigMatrix Class}
\description{
This is the 'mother' class for all
gpuBigMatrix objects.  It directly inherits from
the \link[bigmemory]{big.matrix} class.  All other
gpuBigMatrix classes inherit from this class but
there are no current circumstances where this class
is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{igpuBigMatrix}, \code{fgpuBigMatrix}, and
\code{dgpuBigMatrix} corresponding to integer, float, and
double data types respectively.
}
\section{Slots}{

 Common to all gpuBigMatrix objects in the package
 \describe{
     \item{\code{address}:}{External pointer to shared
     memory matrix.}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{igpuBigMatrix-class}},
\code{\link{fgpuBigMatrix-class}},
\code{\link{dgpuBigMatrix-class}}
}

