% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpuBigMatrix.R
\docType{methods}
\name{gpuBigMatrix}
\alias{gpuBigMatrix}
\alias{gpuBigMatrix,big.matrix}
\alias{gpuBigMatrix,big.matrix-method}
\alias{gpuBigMatrix,matrix}
\alias{gpuBigMatrix,matrix-method}
\title{Construct a gpuBigMatrix}
\usage{
gpuBigMatrix(data = NA, ncol = NA, nrow = NA, type = NULL, ...)

\S4method{gpuBigMatrix}{matrix}(data, ncol = NA, nrow = NA, type = NULL)

\S4method{gpuBigMatrix}{big.matrix}(data, ncol = NA, nrow = NA,
  type = NULL)
}
\arguments{
\item{data}{A matrix object that is or can be converted to a
\code{big.matrix} object}

\item{ncol}{An integer specifying the number of columns}

\item{nrow}{An integer specifying the number of rows}

\item{type}{A character string specifying the type of gpuMatrix.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to gpuBigMatrix methods}
}
\value{
A gpuBigMatrix object
}
\description{
Construct a gpuBigMatrix of a class that inherits
from \code{gpuBigMatrix}
}
\author{
Charles Determan Jr.
}

