% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{crossprod,gpuMatrix,missing-method}
\alias{crossprod,gpuMatrix}
\alias{crossprod,gpuMatrix,gpuMatrix-method}
\alias{crossprod,gpuMatrix,missing-method}
\alias{tcrossprod,gpuMatrix,gpuMatrix-method}
\alias{tcrossprod,gpuMatrix,missing-method}
\title{gpuMatrix Crossproduct}
\usage{
\S4method{crossprod}{gpuMatrix,missing}(x, y)

\S4method{crossprod}{gpuMatrix,gpuMatrix}(x, y)

\S4method{tcrossprod}{gpuMatrix,missing}(x, y)

\S4method{tcrossprod}{gpuMatrix,gpuMatrix}(x, y)
}
\arguments{
\item{x}{A gpuMatrix}

\item{y}{A gpuMatrix}
}
\value{
A gpuMatrix
}
\description{
Return the matrix cross-product of two conformable
matrices using a GPU.  This is equivalent to t(x) %*% y (crossprod)
or x %*% t(t) (tcrossprod) but faster as no data transfer between
device and host is required.
}
\author{
Charles Determan Jr.
}

