% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_matrix.R
\docType{methods}
\name{colSums,gpuMatrix,missing,missing-method}
\alias{colSums,gpuMatrix}
\alias{colSums,gpuMatrix,missing,missing-method}
\alias{rowSums,gpuMatrix,missing,missing-method}
\title{Row and Column Sums of gpuMatrix}
\usage{
\S4method{colSums}{gpuMatrix,missing,missing}(x, na.rm, dims)

\S4method{rowSums}{gpuMatrix,missing,missing}(x, na.rm, dims)
}
\arguments{
\item{x}{A gpuMatrix object}

\item{na.rm}{Not currently used}

\item{dims}{Not currently used}
}
\value{
A gpuVector object
}
\description{
Row and column sums of gpuMatrix objects
}
\author{
Charles Determan Jr.
}

