% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{cov,gpuMatrix,missing,missing,missing-method}
\alias{cov,gpuMatrix}
\alias{cov,gpuMatrix,missing,missing,character-method}
\alias{cov,gpuMatrix,missing,missing,missing-method}
\title{Covariance (gpuMatrix)}
\usage{
\S4method{cov}{gpuMatrix,missing,missing,missing}(x, y = NULL, use = NULL,
  method = "pearson")

\S4method{cov}{gpuMatrix,missing,missing,character}(x, y = NULL, use = NULL,
  method = "pearson")
}
\arguments{
\item{x}{A gpuMatrix object}

\item{y}{Not used}

\item{use}{Not used}

\item{method}{Character string indicating with covariance to be computed.}
}
\value{
A \code{gpuMatrix} containing the symmetric covariance values.
}
\description{
Covariance (gpuMatrix)
}
\author{
Charles Determan Jr.
}

