% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello-gpuR.R
\docType{package}
\name{gpuR-package}
\alias{gpuR-package}
\alias{gpuR}
\title{GPU functions for R Objects}
\description{
This package was developed to provide simple to use R functions
that leverage the power of GPU's but also retain a format familiar to the
R user.  There are a hand full of other R packages that provide some
GPU functionality but nearly all rely on a CUDA backend thereby restricting
the user to NVIDIA GPU hardware.  In the spirit of being as broadly 
applicable as possible, this GPU code herein relies upon OpenCL via the 
ViennaCL library.  

OpenCL, in contrast to CUDA, is open source and can be used across 
different graphics cards (e.g. NVIDIA, AMD, Intel).  This package removes 
the complex code needed for GPU computing and provides easier to use 
functions to apply on R objects.

\tabular{ll}{ Package: \tab gpuR\cr Type: \tab Package\cr 
Version: \tab 1.0.0\cr Date: \tab 2015-03-31\cr License: \tab GPL-3\cr
Copyright: \tab (c) 2015 Charles E. Determan Jr.\cr URL: \tab 
\url{https://github.com/cdeterman/gpuR}\cr LazyLoad: \tab yes\cr
}
}

