% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuSparseMatrix.R

\name{gpuSparseMatrix-class}
\alias{gpuSparseMatrix-class}
\alias{gpuSparseMatrix}
\title{gpuSparseMatrix Class}
\description{
This is the 'mother' class for all gpuSparseMatrix objects. It wraps sparse
matrices in Compressed Sparse Row (CSR) format for efficient GPU computation.
All other gpuSparseMatrix classes inherit from this class.

There are multiple child classes that correspond to the particular data type
contained. These include \code{igpuSparseMatrix}, \code{fgpuSparseMatrix}, and
\code{dgpuSparseMatrix} corresponding to integer, float, and double data types
respectively.
}
\section{Slots}{

Common to all gpuSparseMatrix objects in the package
\describe{
    \item{\code{address}:}{Pointer to sparse matrix data}
    \item{\code{.context_index}:}{Integer index of OpenCL contexts}
    \item{\code{.platform_index}:}{Integer index of OpenCL platforms}
    \item{\code{.platform}:}{Name of OpenCL platform}
    \item{\code{.device_index}:}{Integer index of active device}
    \item{\code{.device}:}{Name of active device}
}
}

