% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuVector.R
\docType{class}
\name{gpuVector-class}
\alias{gpuVector-class}
\title{gpuVector Class}
\description{
This is the 'mother' class for all
gpuVector objects.  All other gpuVector classes
inherit from this class but there are no current
circumstances where this class is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{igpuVector}.
}
\section{Slots}{

 Common to all vclMatrix objects in the package
 \describe{
     \item{\code{address}:}{Pointer to data matrix}
     \item{\code{.context_index}:}{Integer index of OpenCL contexts}
     \item{\code{.platform_index}:}{Integer index of OpenCL platforms}
     \item{\code{.platform}:}{Name of OpenCL platform}
     \item{\code{.device_index}:}{Integer index of active device}
     \item{\code{.device}:}{Name of active device}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{igpuVector-class}}
}

