% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{\%*\%,gpuVector,gpuVector-method}
\alias{\%*\%,gpuVector,gpuVector-method}
\title{gpuVector Inner Product}
\usage{
\S4method{\%*\%}{gpuVector,gpuVector}(x, y)
}
\arguments{
\item{x}{A gpuVector object}

\item{y}{A gpuVector object}
}
\value{
A numeric value
}
\description{
gpuVector Inner Product
}

