% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{\%o\%,gpuVector,gpuVector-method}
\alias{\%o\%,gpuVector,gpuVector-method}
\title{gpuVector Outer Product}
\usage{
\S4method{\%o\%}{gpuVector,gpuVector}(X, Y)
}
\arguments{
\item{X}{A gpuVector object}

\item{Y}{A gpuVector object}
}
\value{
A gpuMatrix object
}
\description{
gpuVector Outer Product
}

