% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R, R/methods-vclVector.R
\docType{methods}
\name{\%o\%,gpuVector,gpuVector-method}
\alias{\%o\%,gpuVector,gpuVector-method}
\alias{\%o\%,vclVector,vclVector-method}
\title{Outer Product}
\usage{
\S4method{\%o\%}{gpuVector,gpuVector}(X, Y)

\S4method{\%o\%}{vclVector,vclVector}(X, Y)
}
\arguments{
\item{X}{A gpuR object}

\item{Y}{A gpuR object}
}
\description{
The outer product of two gpuR vector objects
}
\author{
Charles Determan Jr.
}

