% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{\%o\%,vclVector,vclVector-method}
\alias{\%o\%,vclVector,vclVector-method}
\title{vclVector Outer Product}
\usage{
\S4method{\%o\%}{vclVector,vclVector}(X, Y)
}
\arguments{
\item{X}{A vclVector object}

\item{Y}{A vclVector object}
}
\value{
A vclMatrix object
}

