% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{\%*\%,gpuMatrix,gpuMatrix-method}
\alias{\%*\%,gpuMatrix,gpuMatrix-method}
\title{gpuMatrix Multiplication}
\usage{
\S4method{\%*\%}{gpuMatrix,gpuMatrix}(x, y)
}
\arguments{
\item{x}{A gpuMatrix object}

\item{y}{A gpuMatrix object}
}
\description{
gpuMatrix Multiplication
}

