% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{\%*\%,vclMatrix,vclMatrix-method}
\alias{\%*\%,vclMatrix,vclMatrix-method}
\title{vclMatrix Multiplication}
\usage{
\S4method{\%*\%}{vclMatrix,vclMatrix}(x, y)
}
\arguments{
\item{x}{A vclMatrix object}

\item{y}{A vclMatrix object}
}
\value{
A vclMatrix
}

