% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{has_double_skip}
\alias{has_double_skip}
\title{Skip test for GPU double precision}
\usage{
has_double_skip()
}
\description{
Function to skip testthat tests
if the detected GPU doesn't support double precision
}
