% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{has_multiple_gpu_skip}
\alias{has_multiple_gpu_skip}
\title{Skip test in less than 2 GPUs}
\usage{
has_multiple_gpu_skip()
}
\description{
Function to skip testthat tests
if less than 2 valid GPU's are detected
}

