% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\name{identity_matrix}
\alias{identity_matrix}
\title{Identity Matrix on Device}
\usage{
identity_matrix(x, type = NULL)
}
\arguments{
\item{x}{A numeric value indicating the order of the identity matrix}

\item{type}{A character string specifying the type of gpuMatrix.  Default
is derived from \code{getOption("gpuR.default.type")}.}
}
\value{
A \code{vclMatrix} object
}
\description{
Creates an identity matrix directly on the current device
(e.g. GPU)
}
\note{
This function was only created for \code{vclMatrix} objects as
the copy from CPU to \code{gpuMatrix} is trivial using the base
\link[base]{diag} function.
}
\author{
Charles Determan Jr.
}

