% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{igpuMatrix-class}
\alias{igpuMatrix-class}
\title{igpuMatrix Class}
\description{
An integer type matrix in the S4 \code{gpuMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a integer typed matrix}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{gpuMatrix-class}},
\code{\link{igpuMatrix-class}},
\code{\link{dgpuMatrix-class}}
}

