% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuSparseMatrix.R

\name{igpuSparseMatrix-class}
\alias{igpuSparseMatrix-class}
\alias{igpuSparseMatrix}
\title{igpuSparseMatrix Class}
\description{
An integer type sparse matrix in the S4 \code{gpuSparseMatrix}
representation using Compressed Sparse Row (CSR) format.
}
\section{Slots}{

\describe{
    \item{\code{address}:}{Pointer to an integer typed sparse matrix}
}
}

\value{
If the gpuSparseMatrix object is of type 'integer', returns TRUE, if not, returns an error message.
}

