% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{igpuVector-class}
\alias{igpuVector-class}
\title{igpuVector Class}
\description{
An integer vector in the S4 \code{gpuVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{object}:}{An integer vector object}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{gpuVector-class}}
}

