% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace.R
\docType{methods}
\name{inplace}
\alias{inplace}
\alias{inplace,function,gpuMatrix,gpuMatrix-method}
\alias{inplace,function,gpuVector,gpuVector-method}
\alias{inplace,function,numeric,vclMatrix-method}
\alias{inplace,function,vclMatrix,missing-method}
\alias{inplace,function,vclMatrix,vclMatrix-method}
\alias{inplace,function,vclVector,missing-method}
\alias{inplace,function,vclVector,numeric-method}
\alias{inplace,function,vclVector,vclVector-method}
\title{Inplace Function Wrapper}
\usage{
inplace(f, x, y)

\S4method{inplace}{`function`,vclMatrix,vclMatrix}(f, x, y)

\S4method{inplace}{`function`,vclMatrix,missing}(f, x, y)

\S4method{inplace}{`function`,numeric,vclMatrix}(f, x, y)

\S4method{inplace}{`function`,gpuMatrix,gpuMatrix}(f, x, y)

\S4method{inplace}{`function`,vclVector,vclVector}(f, x, y)

\S4method{inplace}{`function`,vclVector,missing}(f, x, y)

\S4method{inplace}{`function`,vclVector,numeric}(f, x, y)

\S4method{inplace}{`function`,gpuVector,gpuVector}(f, x, y)
}
\arguments{
\item{f}{A function}

\item{x}{A gpuR object}

\item{y}{A gpuR object}
}
\value{
No return, result applied in-place
}
\description{
Applies the provided function in-place on the 
first object passed

#' @rdname inplace-methods
#' @export
setMethod("inplace",
		  signature = c("function", "vclMatrix", "vclVector", "vclVector"),
		  function(f, x, y){
		  	
		  	switch(deparse(substitute(f)),
		  		   `-` = vclMatrix_unary_axpy(x, inplace = TRUE),
		  		   `exp` = vclMatElemExp(x, inplace = TRUE),
		  		   `abs` = vclMatElemAbs(x, inplace = TRUE),
		  		   stop("undefined operation")
		  	)
		  })
}
\author{
Charles Determan Jr.
}

