% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-vclVector.R
\docType{class}
\name{ivclVector-class}
\alias{ivclVector-class}
\title{ivclVector Class}
\description{
An integer vector in the S4 \code{vclVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{An integer vector object}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{vclVector-class}}
}

