% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{log,gpuMatrix-method}
\alias{log,gpuMatrix-method}
\title{gpuMatrix Logarithms}
\usage{
\S4method{log}{gpuMatrix}(x, base = NULL)
}
\arguments{
\item{x}{A gpuMatrix object}

\item{base}{A positive number (complex not currently supported by OpenCL):
the base with respect to which logarithms are computed.  Defaults to the
natural log.}
}
\value{
A gpuMatrix object
}

