% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{log,gpuVector-method}
\alias{log,gpuVector-method}
\title{gpuVector Logarithms}
\usage{
\S4method{log}{gpuVector}(x, base = NULL)
}
\arguments{
\item{x}{A gpuVector object}

\item{base}{A positive number (complex not currently supported by OpenCL):
the base with respect to which logarithms are computed.  Defaults to the
natural log.}
}
\value{
A gpuVector object
}

