% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{log,vclMatrix-method}
\alias{log,vclMatrix-method}
\title{vclMatrix Logarithms}
\usage{
\S4method{log}{vclMatrix}(x, base = NULL)
}
\arguments{
\item{x}{A vclMatrix object}

\item{base}{A positive number (complex not currently supported by OpenCL):
the base with respect to which logarithms are computed.  Defaults to the
natural log.}
}
\value{
A vclMatrix object
}

