% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{log,vclVector-method}
\alias{log,vclVector-method}
\title{vclVector Logarithms}
\usage{
\S4method{log}{vclVector}(x, base = NULL)
}
\arguments{
\item{x}{A vclVector object}

\item{base}{A positive number (complex not currently supported by OpenCL):
the base with respect to which logarithms are computed.  Defaults to the
natural log.}
}
\value{
A vclVector object
}

