% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Math,gpuVector-method}
\alias{Math,gpuVector-method}
\title{gpuVector Math methods}
\usage{
\S4method{Math}{gpuVector}(x)
}
\arguments{
\item{x}{A gpuVector object}
}
\value{
A gpuVector object
}

