% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R, R/methods-vclMatrix.R,
%   R/methods-vclVector.R, R/methods.R
\docType{methods}
\name{Math,gpuVector-method}
\alias{Math,gpuMatrix-method}
\alias{Math,gpuVector-method}
\alias{Math,vclMatrix-method}
\alias{Math,vclVector-method}
\alias{Math-gpuR-method}
\title{gpuR Math methods}
\usage{
\S4method{Math}{gpuVector}(x)

\S4method{Math}{vclMatrix}(x)

\S4method{Math}{vclVector}(x)

\S4method{Math}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuR object}
}
\value{
A gpuR object
}
\description{
Methods for the base Math methods \link[methods]{S4groupGeneric}
}
\details{
Currently implemented methods include:
\itemize{
 \item{"sin", "cos", "tan", "asin", "acos", "atan", "sinh", "cosh", "tanh",
 "log10", "exp", "abs"}
 }
}
\author{
Charles Determan Jr.
}

