% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{Math,vclMatrix-method}
\alias{Math,vclMatrix-method}
\title{vclMatrix Math methods}
\usage{
\S4method{Math}{vclMatrix}(x)
}
\arguments{
\item{x}{A vclMatrix object}
}
\value{
A vclMatrix object
}
\description{
vclMatrix Math methods
}

