% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{Math,vclVector-method}
\alias{Math,vclVector-method}
\title{vclVector Math methods}
\usage{
\S4method{Math}{vclVector}(x)
}
\arguments{
\item{x}{A vclVector object}
}
\value{
A vclVector object
}

