% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_matrix.R
\docType{methods}
\name{ncol,gpuMatrix-method}
\alias{ncol,gpuMatrix-method}
\title{gpuMatrix ncol method}
\usage{
\S4method{ncol}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\description{
gpuMatrix ncol method
}

