% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_matrix.R
\docType{methods}
\name{nrow,gpuMatrix-method}
\alias{nrow,gpuMatrix-method}
\title{gpuMatrix nrow method}
\usage{
\S4method{nrow}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\description{
gpuMatrix nrow method
}

