% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R, R/methods.R
\docType{methods}
\name{nrow,vclMatrix-method}
\alias{ncol,gpuMatrix-method}
\alias{ncol,vclMatrix-method}
\alias{nrow,gpuMatrix-method}
\alias{nrow,vclMatrix-method}
\title{The Number of Rows/Columns of a gpuR matrix}
\usage{
\S4method{nrow}{vclMatrix}(x)

\S4method{ncol}{vclMatrix}(x)

\S4method{nrow}{gpuMatrix}(x)

\S4method{ncol}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix/vclMatrix object}
}
\value{
An integer of length 1
}
\description{
\code{nrow} and \code{ncol} return the number of rows or columns
present in \code{x} respectively.
}
\author{
Charles Determan Jr.
}

