% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{nrow,vclMatrix-method}
\alias{ncol,vclMatrix-method}
\alias{nrow,vclMatrix}
\alias{nrow,vclMatrix-method}
\title{The Number of Rows/Columns of a vclMatrix}
\usage{
\S4method{nrow}{vclMatrix}(x)

\S4method{ncol}{vclMatrix}(x)
}
\arguments{
\item{x}{A vclMatrix object}
}
\value{
An integer of length 1
}
\author{
Charles Determan Jr.
}

