% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{permute}
\alias{permute}
\title{Permuting functions for \code{gpuR} objects}
\usage{
permute(x, MARGIN, order)
}
\arguments{
\item{x}{A \code{gpuR} matrix object}

\item{MARGIN}{dimension over which the ordering should be applied, 1
indicates rows, 2 indicates columns}

\item{order}{An integer vector indicating order of rows to assign}
}
\value{
A \code{gpuR} object
}
\description{
Generate a permutation of row or column indices
}
\author{
Charles Determan Jr.
}
