% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpuSparseMatrix.R

\name{print-gpuSparseMatrix}
\alias{print,gpuSparseMatrix-method}
\title{Print GPU Sparse Matrix}
\usage{
\S4method{print}{gpuSparseMatrix}(x)
}
\arguments{
\item{x}{A gpuSparseMatrix object}
}
\description{
Print method for GPU sparse matrices displaying basic properties
}

