% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\name{qr.Q,gpuQR-method}
\alias{qr.Q,gpuQR-method}
\title{Reconstruct the Q Matrix from a gpuQR Object}
\usage{
\S4method{qr.Q}{gpuQR}(qr, complete = FALSE)
}
\arguments{
\item{qr}{\code{gpuQR} object}

\item{complete}{not currently used}
}
\value{
\code{qr.Q} returns all of \code{Q},
}
\description{
Returns the components of the QR decomposition.
}
\seealso{
\link[base]{qr.R}, \link[base]{qr.Q}
}
\author{
Charles Determan Jr.
}
