% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr.R
\docType{methods}
\name{qr.R,gpuQR-method}
\alias{qr.R,gpuQR-method}
\alias{qr.Q,gpuQR-method}
\title{Reconstruct the Q or R Matrices form a gpuQR Object}
\usage{
\S4method{qr.R}{gpuQR}(qr, complete = FALSE)

\S4method{qr.Q}{gpuQR}(qr, complete = FALSE)
}
\arguments{
\item{qr}{\code{gpuQR} object}

\item{complete}{not currently used}
}
\value{
\code{qr.Q} returns all of \code{Q}
\code{qr.R} returns all of \code{R}
}
\description{
Returns the components of the QR decomposition.
}
\seealso{
\link[base]{qr.R}, \link[base]{qr.Q}
}
\author{
Charles Determan Jr.
}
