% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{set_device_context}
\alias{set_device_context}
\title{Set Context for Specific Device Type}
\usage{
set_device_context(type)
}
\arguments{
\item{type}{A character vector specifying device type}
}
\value{
An integer indicating previous context index
}
\description{
This function find the first context
that contains a device of the specified type.
}
