% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers-utils.R
\name{setContext}
\alias{setContext}
\title{Set Context}
\usage{
setContext(id = 1L)
}
\arguments{
\item{id}{Integer identifying which context to set}
}
\description{
Change the current context used by default
}
\seealso{
\link{listContexts}
}

