% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\docType{methods}
\name{solve,vclMatrix,vclMatrix-method}
\alias{solve,gpuMatrix,gpuMatrix-method}
\alias{solve,gpuMatrix,missing-method}
\alias{solve,vclMatrix,missing-method}
\alias{solve,vclMatrix,vclMatrix-method}
\title{Solve a System of Equations for gpuR objects}
\usage{
\S4method{solve}{vclMatrix,vclMatrix}(a, b, ...)

\S4method{solve}{vclMatrix,missing}(a, b, ...)

\S4method{solve}{gpuMatrix,gpuMatrix}(a, b, ...)

\S4method{solve}{gpuMatrix,missing}(a, b, ...)
}
\arguments{
\item{a}{A gpuR object}

\item{b}{A gpuR object}

\item{...}{further arguments passed to or from other methods}
}
\value{
A gpuR object
}
\description{
This funciton solves the equation \code{a \%*\% x = b} for 
\code{x}, where \code{b} can be either a vector or a matrix.
}
\author{
Charles Determan Jr.
}

