% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{[,gpuMatrix,missing,missing,missing-method}
\alias{[,gpuMatrix}
\alias{[,gpuMatrix,missing,missing,missing-method}
\title{Extract all gpuMatrix elements}
\usage{
\S4method{[}{gpuMatrix,missing,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A gpuMatrix object}

\item{i}{missing}

\item{j}{missing}

\item{drop}{missing}
}
\description{
Extract all gpuMatrix elements
}
\author{
Charles Determan Jr.
}

