% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{[,vclMatrix,missing,missing,missing-method}
\alias{[,vclMatrix}
\alias{[,vclMatrix,missing,missing,missing-method}
\title{Extract all vclMatrix elements}
\usage{
\S4method{[}{vclMatrix,missing,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A vclMatrix object}

\item{i}{missing}

\item{j}{missing}

\item{drop}{missing}
}
\description{
Extract all vclMatrix elements
}
\author{
Charles Determan Jr.
}

