% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{[,vclVector,missing,missing,missing-method}
\alias{[,vclVector}
\alias{[,vclVector,missing,missing,missing-method}
\title{Extract all vclVector elements}
\usage{
\S4method{[}{vclVector,missing,missing,missing}(x, i, j, drop)
}
\arguments{
\item{x}{A vclVector object}

\item{i}{missing}

\item{j}{missing}

\item{drop}{missing}
}
\description{
Extract all vclVector elements
}
\author{
Charles Determan Jr.
}

