% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{Summary,gpuMatrix-method}
\alias{Summary,gpuMatrix-method}
\title{gpuMatrix Summary methods}
\usage{
\S4method{Summary}{gpuMatrix}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\value{
For \code{min} or \code{max}, a length-one vector
}

