% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{Summary,gpuVector-method}
\alias{Summary,gpuVector-method}
\title{gpuVector Summary methods}
\usage{
\S4method{Summary}{gpuVector}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A gpuVector object}
}
\value{
For \code{min} or \code{max}, a length-one vector
}

