% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{Summary,vclMatrix-method}
\alias{Summary,vclMatrix-method}
\title{vclMatrix Summary methods}
\usage{
\S4method{Summary}{vclMatrix}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A vclMatrix object}
}
\value{
For \code{min} or \code{max}, a length-one vector
}

