% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{Summary,vclVector-method}
\alias{Summary,vclVector-method}
\title{vclVector Summary methods}
\usage{
\S4method{Summary}{vclVector}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A vclVector object}
}
\value{
For \code{min} or \code{max}, a length-one vector
}

