% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{synchronize}
\alias{synchronize}
\title{Synchronize Device Execution}
\usage{
synchronize()
}
\value{
This does not return anything.
}
\description{
This pauses execution until the processing is complete
on the device (CPU/GPU/etc.).  This is important especially for
benchmarking applications.
}

