% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R, R/methods.R
\docType{methods}
\name{t,vclMatrix-method}
\alias{t,gpuMatrix}
\alias{t,gpuMatrix-method}
\alias{t,vclMatrix}
\alias{t,vclMatrix-method}
\title{\code{gpuR} matrix transpose}
\usage{
\S4method{t}{vclMatrix}(x)

\S4method{t}{gpuMatrix}(x)
}
\arguments{
\item{x}{A \code{gpuR} matrix}
}
\value{
A \code{gpuR} matrix
}
\description{
Given a \code{gpuR} matrix \code{x}, \code{t} returns the 
transpose of \code{x}
}
\author{
Charles Determan Jr.
}

