% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof.R
\docType{methods}
\name{typeof,gpuMatrix-method}
\alias{typeof,gpuMatrix}
\alias{typeof,gpuMatrix-method}
\title{Get gpuMatrix type}
\usage{
\S4method{typeof}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}

