% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof.R
\docType{methods}
\name{typeof,gpuMatrix-method}
\alias{typeof,gpuMatrix-method}
\alias{typeof,gpuVector-method}
\alias{typeof,vclMatrix-method}
\alias{typeof,vclVector-method}
\title{Get gpuR object type}
\usage{
\S4method{typeof}{gpuMatrix}(x)

\S4method{typeof}{gpuVector}(x)

\S4method{typeof}{vclMatrix}(x)

\S4method{typeof}{vclVector}(x)
}
\arguments{
\item{x}{A gpuR object}
}
\description{
\code{typeof} determines the type (i.e. storage mode) of a 
gpuR object
}
\author{
Charles Determan Jr.
}
