% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{typeof,gpuVector-method}
\alias{typeof,gpuVector}
\alias{typeof,gpuVector-method}
\title{Get gpuVector type}
\usage{
\S4method{typeof}{gpuVector}(x)
}
\arguments{
\item{x}{A gpuVector object}
}
\description{
Get gpuVector type
}

