% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVectorSlice.R
\docType{methods}
\name{typeof,gpuVectorSlice-method}
\alias{typeof,gpuVectorSlice}
\alias{typeof,gpuVectorSlice-method}
\title{Get gpuVectorSlice type}
\usage{
\S4method{typeof}{gpuVectorSlice}(x)
}
\arguments{
\item{x}{A gpuVectorSlice object}
}

