% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{typeof,vclVector-method}
\alias{typeof,vclVector}
\alias{typeof,vclVector-method}
\title{Get vclVector type}
\usage{
\S4method{typeof}{vclVector}(x)
}
\arguments{
\item{x}{A vclVector object}
}

