% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vclMatrix.R
\docType{methods}
\name{vclMatrix}
\alias{vclMatrix}
\alias{vclMatrix,matrix}
\alias{vclMatrix,matrix-method}
\alias{vclMatrix,missing}
\alias{vclMatrix,missing-method}
\title{Construct a vclMatrix}
\usage{
vclMatrix(data = NA, nrow = NA, ncol = NA, type = NULL, ...)

\S4method{vclMatrix}{matrix}(data, type = NULL)

\S4method{vclMatrix}{missing}(data, nrow = NA, ncol = NA, type = NULL)
}
\arguments{
\item{data}{An object that is or can be converted to a
\code{matrix} object}

\item{nrow}{An integer specifying the number of rows}

\item{ncol}{An integer specifying the number of columns}

\item{type}{A character string specifying the type of vclMatrix.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to vclMatrix methods}
}
\value{
A vclMatrix object
}
\description{
Construct a vclMatrix of a class that inherits
from \code{vclMatrix}
}
\author{
Charles Determan Jr.
}

