% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{cov,vclMatrix,missing,missing,missing-method}
\alias{cov,vclMatrix}
\alias{cov,vclMatrix,missing,missing,character-method}
\alias{cov,vclMatrix,missing,missing,missing-method}
\title{Covariance (vclMatrix)}
\usage{
\S4method{cov}{vclMatrix,missing,missing,missing}(x, y = NULL, use = NULL,
  method = "pearson")

\S4method{cov}{vclMatrix,missing,missing,character}(x, y = NULL, use = NULL,
  method = "pearson")
}
\arguments{
\item{x}{A vclMatrix object}

\item{y}{Not used}

\item{use}{Not used}

\item{method}{Character string indicating which covariance to be computed.}
}
\value{
A \code{vclMatrix} containing the symmetric covariance values.
}
\description{
Covariance (vclMatrix)
}
\author{
Charles Determan Jr.
}

