% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclVector.R
\docType{class}
\name{vclVector-class}
\alias{vclVector-class}
\title{vclVector Class}
\description{
This is the 'mother' class for all
vclVector objects.  All other vclVector classes
inherit from this class but there are no current
circumstances where this class is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{ivclVector}.
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{ivclVector-class}}
}

