% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuSparseMatrix.R

\name{zgpuSparseMatrix-class}
\alias{zgpuSparseMatrix-class}
\alias{zgpuSparseMatrix}
\title{zgpuSparseMatrix Class}
\description{
A complex double type sparse matrix in the S4 \code{gpuSparseMatrix}
representation using Compressed Sparse Row (CSR) format.
}
\section{Slots}{

\describe{
    \item{\code{address}:}{Pointer to a complex double sparse matrix.}
}
}

\value{
If the gpuSparseMatrix object is of type 'complex double', returns TRUE, if not, returns an error message.
}

