% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclMatrix.R
\docType{class}
\name{zvclMatrix-class}
\alias{zvclMatrix-class}
\title{zvclMatrix Class}
\description{
An complex double type matrix in the S4 \code{vclMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a complex double type matrix}
 }
}

\seealso{
\code{\link{vclMatrix-class}}, 
\code{\link{ivclMatrix-class}},
\code{\link{fvclMatrix-class}}
}
\author{
Charles Determan Jr.
}
