% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers-utils.R
\name{detectGPUs}
\alias{detectGPUs}
\title{Detect Available GPUs}
\usage{
detectGPUs(platform_idx = NULL)
}
\arguments{
\item{platform_idx}{An integer value indicating which platform to query.
If NULL it will iterate over all platforms and sum results}
}
\value{
An integer representing the number of available GPUs
}
\description{
Find out how many GPUs available
}
\seealso{
\link{detectPlatforms}
}

