% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers-utils.R
\name{gpuInfo}
\alias{cpuInfo}
\alias{gpuInfo}
\title{Device Information}
\usage{
gpuInfo(platform_idx = NULL, device_idx = NULL)

cpuInfo(platform_idx = NULL, device_idx = NULL)
}
\arguments{
\item{platform_idx}{An integer value indicating which platform to query.}

\item{device_idx}{An integer value indicating which device to query.}
}
\value{
\item{deviceName}{Device Name}

\item{deviceVendor}{Device Vendor}

\item{numberOfCores}{Number of Computing Units 
 (which execute the work groups)}

\item{maxWorkGroupSize}{Maximum number of work items
 per group}

\item{maxWorkItemDim}{Number of dimensions}

\item{maxWorkItemSizes}{Maximum number of works items
 per dimension}

\item{deviceMemory}{Global amount of memory (bytes)}

\item{clockFreq}{Maximum configured clock frequency of the 
device in MHz}

\item{localMem}{Maximum amount of local memory for each work 
group (bytes)}

\item{maxAllocatableMem}{Maximum amount of memory in a single 
piece (bytes)}

\item{available}{Whether the device is available}

\item{deviceExtensions}{OpenCL device extensions available}

\item{double_support}{Logical value if double type supported}
}
\description{
Get basic information about selected device (e.g. GPU)
}
\author{
Charles Determan Jr.
}
\seealso{
\link{detectPlatforms} \link{detectGPUs} \link{detectCPUs} \link{cpuInfo}
}

