% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers-utils.R
\name{deviceType}
\alias{deviceType}
\title{Check device type}
\usage{
deviceType(platform_idx = 1L, device_idx = 1L)
}
\arguments{
\item{platform_idx}{An integer value indicating which platform to query.}

\item{device_idx}{An integer value indicating which device to query.}
}
\value{
A character string indicating the device type
}
\description{
Check what type a device is given platform and device indices
}

