% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class-gpuMatrix.R
\docType{class}
\name{dgpuMatrix-class}
\alias{dgpuMatrix-class}
\title{dgpuMatrix Class}
\description{
An integer type matrix in the S4 \code{gpuMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double type matrix}
 }
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{gpuMatrix-class}},
\code{\link{igpuMatrix-class}},
\code{\link{fgpuMatrix-class}}
}

