% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuMatrix.R
\docType{class}
\name{dgpuMatrix-class}
\alias{dgpuMatrix-class}
\title{dgpuMatrix Class}
\description{
A double type matrix in the S4 \code{gpuMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double type matrix}
 }
}

\seealso{
\code{\link{gpuMatrix-class}}, 
\code{\link{igpuMatrix-class}},
\code{\link{fgpuMatrix-class}}
}
\author{
Charles Determan Jr.
}
