% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuVector.R
\docType{class}
\name{dgpuVector-class}
\alias{dgpuVector-class}
\title{dgpuVector Class}
\value{
If the gpuVector object is of type 'double', returns TRUE, if not, returns an error message.
}
\description{
An double vector in the S4 \code{gpuVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double typed vector}
 }
}

\seealso{
\code{\link{gpuVector-class}}
}
\author{
Charles Determan Jr.
}
