% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{dim,gpuMatrix-method}
\alias{dim,gpuMatrix}
\alias{dim,gpuMatrix-method}
\title{gpuMatrix dim method}
\usage{
\S4method{dim}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\value{
A length 2 vector of the number of rows and columns respectively.
}
\description{
gpuMatrix dim method
}
\author{
Charles Determan Jr.
}

