% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R, R/methods.R
\docType{methods}
\name{dim,vclMatrix-method}
\alias{dim,gpuMatrix-method}
\alias{dim,vclMatrix-method}
\alias{dim-gpuMatrix}
\alias{dim-vclMatrix}
\title{gpuMatrix/vclMatrix dim method}
\usage{
\S4method{dim}{vclMatrix}(x)

\S4method{dim}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix/vclMatrix object}
}
\value{
A length 2 vector of the number of rows and columns respectively.
}
\description{
Retrieve dimension of object
}
\author{
Charles Determan Jr.
}

