% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclVector.R
\docType{class}
\name{dvclVector-class}
\alias{dvclVector-class}
\title{dvclVector Class}
\value{
If the vclVector object is of type 'double', returns TRUE, if not, returns an error message.
}
\description{
An double vector in the S4 \code{vclVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a double typed vector}
 }
}

\seealso{
\code{\link{vclVector-class}}
}
\author{
Charles Determan Jr.
}
