% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{[,gpuMatrix,missing,missing,missing-method}
\alias{[,gpuMatrix}
\alias{[,gpuMatrix,missing,missing,missing-method}
\alias{[,gpuMatrix,missing,numeric,missing-method}
\alias{[,gpuMatrix,numeric,missing,missing-method}
\alias{[,gpuMatrix,numeric,numeric,missing-method}
\alias{[<-,gpuMatrix,missing,numeric,numeric-method}
\alias{[<-,gpuMatrix,numeric,missing,numeric-method}
\alias{[<-,gpuMatrix,numeric,numeric,numeric-method}
\alias{[<-,igpuMatrix,missing,numeric,integer-method}
\alias{[<-,igpuMatrix,numeric,missing,integer-method}
\alias{[<-,igpuMatrix,numeric,numeric,integer-method}
\title{Extract gpuMatrix elements}
\usage{
\S4method{[}{gpuMatrix,missing,missing,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,missing,numeric,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,numeric,missing,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,numeric,numeric,missing}(x, i, j, drop)

\S4method{[}{gpuMatrix,numeric,missing,numeric}(x, i, j) <- value

\S4method{[}{igpuMatrix,numeric,missing,integer}(x, i, j) <- value

\S4method{[}{gpuMatrix,missing,numeric,numeric}(x, i, j) <- value

\S4method{[}{igpuMatrix,missing,numeric,integer}(x, i, j) <- value

\S4method{[}{gpuMatrix,numeric,numeric,numeric}(x, i, j) <- value

\S4method{[}{igpuMatrix,numeric,numeric,integer}(x, i, j) <- value
}
\arguments{
\item{x}{A gpuMatrix object}

\item{i}{indices specifying rows}

\item{j}{indices specifying columns}

\item{drop}{missing}

\item{value}{data of similar type to be added to gpuMatrix object}
}
\author{
Charles Determan Jr.
}

