% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclMatrix.R
\docType{class}
\name{fvclMatrix-class}
\alias{fvclMatrix-class}
\title{fvclMatrix Class}
\value{
If the vclMatrix object is of type 'float', returns TRUE, if not, returns an error message.
}
\description{
An integer type matrix in the S4 \code{vclMatrix}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{Pointer to a float matrix.}
 }
}

\seealso{
\code{\link{vclMatrix-class}}, 
\code{\link{ivclMatrix-class}},
\code{\link{dvclMatrix-class}}
}
\author{
Charles Determan Jr.
}
