% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gpuVector-class}
\alias{gpuVector-class}
\title{gpuVector Class}
\description{
This is the 'mother' class for all
gpuVector objects.  All other gpuVector classes
inherit from this class but there are no current
circumstances where this class is used directly.

There are multiple child classes that correspond
to the particular data type contained.  These include
\code{igpuVector}.
}
\author{
Charles Determan Jr.
}
\seealso{
\code{\link{igpuVector-class}}
}

