% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpuVector.R
\docType{methods}
\name{gpuVector}
\alias{gpuVector}
\alias{gpuVector,missingOrNULL}
\alias{gpuVector,missingOrNULL,ANY-method}
\alias{gpuVector,vector}
\alias{gpuVector,vector,missing-method}
\title{Construct a gpuVector}
\usage{
gpuVector(data, length, type = NULL, ...)

\S4method{gpuVector}{vector,missing}(data, length, type = NULL)

\S4method{gpuVector}{missingOrNULL,ANY}(data, length, type = NULL)
}
\arguments{
\item{data}{An object that is or can be converted to a
\code{vector}}

\item{length}{A non-negative integer specifying the desired length.}

\item{type}{A character string specifying the type of gpuVector.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to gpuVector methods}
}
\value{
A gpuVector object
}
\description{
Construct a gpuVector of a class that inherits
from \code{gpuVector}
}
\author{
Charles Determan Jr.
}

