% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-vclVector.R
\docType{methods}
\name{\%*\%,vclVector,vclVector-method}
\alias{\%*\%,vclVector,vclVector-method}
\title{vclVector Dot Product}
\usage{
\S4method{\%*\%}{vclVector,vclVector}(x, y)
}
\arguments{
\item{x}{A vclVector object}

\item{y}{A vclVector object}
}
\value{
A vclVector
}
\description{
vclVector Dot Product
}

