% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{has_multiple_double_skip}
\alias{has_multiple_double_skip}
\title{Skip test for multiple GPUs with double precision}
\usage{
has_multiple_double_skip()
}
\description{
Function to skip testthat tests
if their aren't multiple detected GPU with double precision
}

