% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-gpuVector.R
\docType{class}
\name{igpuVector-class}
\alias{igpuVector-class}
\title{igpuVector Class}
\value{
If the gpuVector object is of type 'integer', returns TRUE, if not, returns an error message.
}
\description{
An integer vector in the S4 \code{gpuVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{An integer vector object}
 }
}

\seealso{
\code{\link{gpuVector-class}}
}
\author{
Charles Determan Jr.
}
