% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-vclVector.R
\docType{class}
\name{ivclVector-class}
\alias{ivclVector-class}
\title{ivclVector Class}
\value{
If the vclVector object is of type 'integer', returns TRUE, if not, returns an error message.
}
\description{
An integer vector in the S4 \code{vclVector}
representation.
}
\section{Slots}{

 \describe{
     \item{\code{address}:}{An integer vector object}
 }
}

\seealso{
\code{\link{vclVector-class}}
}
\author{
Charles Determan Jr.
}
