% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuVector.R, R/methods-vclVector.R
\docType{methods}
\name{length,gpuVector-method}
\alias{length,gpuVector}
\alias{length,gpuVector-method}
\alias{length,vclVector}
\alias{length,vclVector-method}
\title{Length of gpuVector}
\usage{
\S4method{length}{gpuVector}(x)

\S4method{length}{vclVector}(x)
}
\arguments{
\item{x}{A gpuVector or vclVector object}
}
\value{
A numeric value
}

