% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods-gpuVector.R
\docType{methods}
\name{length,gpuVector-method}
\alias{length,gpuVector}
\alias{length,gpuVector-method}
\title{Length of gpuVector}
\usage{
\S4method{length}{gpuVector}(x)
}
\arguments{
\item{x}{A gpuVector object}
}
\description{
Length of gpuVector
}

