% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-gpuMatrix.R
\docType{methods}
\name{Math,gpuMatrix-method}
\alias{Math,gpuMatrix-method}
\title{gpuMatrix Math methods}
\usage{
\S4method{Math}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\value{
A gpuMatrix object
}

