% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\alias{norm.vclMatrix}
\alias{norm.gpuMatrix}
\title{Compute the Norm of a Matrix}
\usage{
norm(x, type = "O")

\method{norm}{vclMatrix}(x, type = "O")

\method{norm}{gpuMatrix}(x, type = "O")
}
\seealso{
\link[base]{norm}
}
\author{
Charles Determan Jr.
}
