% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax.R
\name{pmax}
\alias{pmax}
\alias{pmin}
\alias{pmin.vclVector}
\title{Parallel Maxima and Minima}
\usage{
pmax(...)

pmin(...)

\method{pmin}{vclVector}(..., na.rm = FALSE)
}
\arguments{
\item{...}{a vclVector object}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A vclMatrix object.
}
\description{
\code{pmax} and \code{pmin} take one or more
vectors as arguments and return a single vector giving the 'parallel'
maxima (or minima) of the argument vectors
}
\seealso{
\link[base]{pmax} \link[base]{pmin}
}
