% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmax.R
\name{pmin.vclVector}
\alias{pmin.vclVector}
\title{# @rdname
setGeneric("pmax", signature = "...")}
\usage{
\method{pmin}{vclVector}(..., na.rm = FALSE)
}
\arguments{
\item{...}{a vclVector object}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\description{
# @rdname 
#' @method pmax vclVector
#' @export
setMethod("pmax",
          "vclVector",
          function(..., na.rm = FALSE){
              print("called correctly")
              gpuVecpmax(..., order = 1L)
          })
}
\details{
#' @export
setMethod("pmax",
          "vclMatrix",
          function(..., na.rm = FALSE){
              gpuMatpmax(..., order = 1L)
          })

#' @export
setMethod("pmax",
          "gpuVector",
          function(..., na.rm = FALSE){
              gpuVecpmax(..., order = 1L)
          })

#' @export
setMethod("pmax",
          "gpuMatrix",
          function(..., na.rm = FALSE){
              gpuMatpmax(..., order = 1L)
          })
}
