% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_utils.R
\name{pocl_check}
\alias{pocl_check}
\title{POCL Version Check}
\usage{
pocl_check()
}
\description{
Versions of POCL up to 0.15-pre have a bug
which results in values being returned when NA values
should be (e.g. fractional powers of negative values)
}
