% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gpuMatrix}
\alias{print.gpuMatrix}
\title{S3 print for gpuMatrix objects}
\usage{
\method{print}{gpuMatrix}(x, ..., n = NULL, width = NULL)
}
\arguments{
\item{x}{A gpuMatrix object}

\item{...}{Additional arguments to print}

\item{n}{Number of rows to display}

\item{width}{Number of columns to display}
}
\description{
prints a gpuMatrix object that is truncated to fit the screen
}

