% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{synchronize}
\alias{synchronize}
\title{Synchronize Device Execution}
\usage{
synchronize()
}
\description{
This pauses execution until the processing is complete
on the device (CPU/GPU/etc.).  This is important especially for
benchmarking applications.
}
\examples{
\dontrun{
    mat <- vclMatrix(rnorm(500^2), ncol = 500, nrow = 500)
    system.time(mat \%*\% mat)
    system.time(mat \%*\% mat; synchronize())
}
}
\author{
Charles Determan Jr.
}
