% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/methods_matrix.R
\docType{methods}
\name{typeof,gpuMatrix-method}
\alias{typeof,gpuMatrix}
\alias{typeof,gpuMatrix-method}
\title{Get gpuMatrix type}
\usage{
\S4method{typeof}{gpuMatrix}(x)
}
\arguments{
\item{x}{A gpuMatrix object}
}
\description{
Get gpuMatrix type
}

