% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeof.R
\docType{methods}
\name{typeof,vclMatrix-method}
\alias{typeof,vclMatrix}
\alias{typeof,vclMatrix-method}
\title{Get vclMatrix type}
\usage{
\S4method{typeof}{vclMatrix}(x)
}
\arguments{
\item{x}{A vclMatrix object}
}

