% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vclMatrix.R
\docType{methods}
\name{vclMatrix}
\alias{vclMatrix}
\alias{vclMatrix,integer}
\alias{vclMatrix,integer-method}
\alias{vclMatrix,matrix}
\alias{vclMatrix,matrix-method}
\alias{vclMatrix,missing}
\alias{vclMatrix,missing-method}
\alias{vclMatrix,numeric-method}
\alias{vclMatrix,vector}
\title{Construct a vclMatrix}
\usage{
vclMatrix(data = NA, nrow = NA, ncol = NA, type = NULL, ...)

\S4method{vclMatrix}{matrix}(data, type = NULL, ctx_id = NULL)

\S4method{vclMatrix}{missing}(data, nrow = NA, ncol = NA, type = NULL,
  ctx_id = NULL)

\S4method{vclMatrix}{numeric}(data, nrow, ncol, type = NULL, ctx_id = NULL)

\S4method{vclMatrix}{integer}(data, nrow, ncol, type = NULL, ctx_id = NULL)
}
\arguments{
\item{data}{An object that is or can be converted to a 
\code{matrix} object}

\item{nrow}{An integer specifying the number of rows}

\item{ncol}{An integer specifying the number of columns}

\item{type}{A character string specifying the type of vclMatrix.  Default
is NULL where type is inherited from the source data type.}

\item{...}{Additional method to pass to vclMatrix methods}

\item{ctx_id}{An integer specifying the object's context}
}
\value{
A vclMatrix object
}
\description{
Construct a vclMatrix of a class that inherits
from \code{vclMatrix}.  This class points to memory directly on
the GPU to avoid the cost of data transfer between host and device.
}
\author{
Charles Determan Jr.
}

