% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{colSums,vclMatrix,missing,missing-method}
\alias{colSums,vclMatrix,missing,missing-method}
\alias{colSums,vclMatrix}
\alias{rowSums,vclMatrix,missing,missing-method}
\alias{colMeans,vclMatrix,missing,missing-method}
\alias{rowMeans,vclMatrix,missing,missing-method}
\title{Row and Column Sums and Means of vclMatrix}
\usage{
\S4method{colSums}{vclMatrix,missing,missing}(x, na.rm, dims)

\S4method{rowSums}{vclMatrix,missing,missing}(x, na.rm, dims)

\S4method{colMeans}{vclMatrix,missing,missing}(x, na.rm, dims)

\S4method{rowMeans}{vclMatrix,missing,missing}(x, na.rm, dims)
}
\arguments{
\item{x}{A vclMatrix object}

\item{na.rm}{Not currently used}

\item{dims}{Not currently used}
}
\value{
A gpuVector object
}
\description{
Row and column sums and of vclMatrix objects
}
\author{
Charles Determan Jr.
}
