% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-vclMatrix.R
\docType{methods}
\name{colSums,vclMatrix-method}
\alias{colSums,vclMatrix-method}
\alias{colSums,vclMatrix}
\alias{rowSums,vclMatrix}
\alias{rowSums,vclMatrix-method}
\alias{colMeans,vclMatrix-method}
\alias{rowMeans,vclMatrix-method}
\title{Row and Column Sums and Means of \code{vclMatrix}}
\usage{
\S4method{colSums}{vclMatrix}(x)

\S4method{rowSums}{vclMatrix}(x)

\S4method{colMeans}{vclMatrix}(x)

\S4method{rowMeans}{vclMatrix}(x)
}
\arguments{
\item{x}{A \code{vclMatrix} object}
}
\value{
A \code{vclVector} object
}
\description{
Row and column sums and of vclMatrix objects
}
\author{
Charles Determan Jr.
}
